backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pydantic-settings>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
PyJWT[crypto]>=2.8.0
httpx>=0.27.0
python-multipart>=0.0.9
google-genai>=1.0.0
markdown-it-py>=3.0.0
